updated Kafka interfaces.
